added is_key_expiring and get_expiring_keys to key_rotation; added a handler for listing out all expiring keys

+++ data_steward/admin/key_rotation.py
@@ -7,6 +7,7 @@ import logging
 LOGGER = logging.getLogger(__name__)
 
 KEY_EXPIRE_DAYS = 180
+KEY_EXPIRING_DAY_THRESHOLD = 170
 
 
 def get_iam_service():
@@ -48,10 +49,24 @@ def is_key_expired(_key):
     :param _key: service account key object
     :return: True if the key exceeds the expiration period, False otherwise
     """
+
     today_date = datetime.today().date()
     created_date = datetime.strptime(_key['validAfterTime'], '%Y-%m-%dT%H:%M:%SZ').date()
     delta = today_date - created_date
     return delta.days > KEY_EXPIRE_DAYS
+
+
+def is_key_expiring(_key):
+    """
+    Determine if a key is expiring soon based on the global field KEY_EXPIRING_DAY_THRESHOLD
+
+    :param _key: service account key object
+    :return: True if the key is expiring soon based, False otherwise
+    """
+    today_date = datetime.today().date()
+    created_date = datetime.strptime(_key['validAfterTime'], '%Y-%m-%dT%H:%M:%SZ').date()
+    delta = today_date - created_date
+    return KEY_EXPIRING_DAY_THRESHOLD <= delta.days <= KEY_EXPIRE_DAYS
 
 
 def delete_key(_key):
@@ -77,12 +92,37 @@ def delete_expired_keys(project_id):
     Delete all expired service account keys associated with a project
 
     :param project_id: identifies the project
-    :return:
+    :return: a list of dicts that contain the info about the deleted keys
     """
+    _deleted_keys = []
     for _service_account in list_service_accounts(project_id):
-        for key in list_keys_for_service_account(_service_account['email']):
-            if is_key_expired(key):
-                delete_key(key)
+        for _key in list_keys_for_service_account(_service_account['email']):
+            if is_key_expired(_key):
+                delete_key(_key)
+
+                _deleted_keys.append({'service_account_email': _service_account['email'],
+                                      'key_name': _key['name'],
+                                      'created_at': _key['validAfterTime']})
+
+    return _deleted_keys
+
+
+def get_expiring_keys(project_id):
+    """
+    List all expiring service account keys associated with a project
+
+    :param project_id: identifies the project
+    :return: a list of dicts that contain the info about the expiring keys
+    """
+    _expiring_keys = []
+    for _service_account in list_service_accounts(project_id):
+        for _key in list_keys_for_service_account(_service_account['email']):
+            if is_key_expiring(_key):
+                _expiring_keys.append({'service_account_email': _service_account['email'],
+                                       'key_name': _key['name'],
+                                       'created_at': _key['validAfterTime']})
+
+    return _expiring_keys
 
 
 if __name__ == '__main__':
@@ -91,6 +131,18 @@ if __name__ == '__main__':
     PARSER = argparse.ArgumentParser(
         description='Delete all expired service account keys associated with a project',
         formatter_class=argparse.RawDescriptionHelpFormatter)
-    PARSER.add_argument('project_id', help='Identifies the project')
+    PARSER.add_argument('-p', '--project_id',
+                        help='Identifies the project',
+                        dest='project_id',
+                        required=True)
+    PARSER.add_argument('-d', '--delete',
+                        help='A flag to indicate whether or not to delete the keys',
+                        action='store_true')
     ARGS = PARSER.parse_args()
-    delete_expired_keys(ARGS.project_id)
+
+    if ARGS.delete:
+        deleted_keys = delete_expired_keys(ARGS.project_id)
+    else:
+        expiring_keys = get_expiring_keys(ARGS.project_id)
+        for key in expiring_keys:
+            print(key)
